# backend/src/types/domain.ts
export type PerfilUsuario = "FORNECEDOR" | "FISCAL_EXTERNO" | "FISCAL_SUAPE" | "ADMIN";

export type StatusRdo =
  | "RASCUNHO"
  | "ENVIADO"
  | "APROVADO_EXTERNO"
  | "APROVADO_SUAPE"
  | "REPROVADO";

export type TipoMidia = "FOTO" | "VIDEO" | "ANEXO";

export type StatusAprovacao = "APROVADO" | "REPROVADO" | "DEVOLVIDO";